requirements.txt
asgiref==3.7.2
Django==4.2.3
django-debug-toolbar==4.1.0
django-tinymce==3.6.1
docopt==0.6.2
install==1.3.5
num2words==0.5.12
Pillow==10.0.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
typing_extensions==4.7.1
psycopg2-binary==2.9.6
celery==5.3.1
redis==4.6.0
django-celery-beat==2.5.0
